GG-48116 Move IgniteUtils.downloadUrl to GridTestUtils
